Fix for signals check
:Release Notes:
Remove signals check from CMakeLists

:Detailed Notes:
Deprecated signals library was dropped in 1.69.0 release:
https://www.boost.org/users/history/version_1_69_0.html
Signals (v1) is now removed. Its removal was announced in 1.68 and its
deprecation was announced in 1.54. Boost 1.68 is the last release that
provides this library. Users are encouraged to use Signals2 instead. The
Boost community thanks Douglas Gregor for his work on Signals which
served its users well and which also inspired Signals2.
This component is already using Signals2 so just drop the check for old
unused component.

:Testing Performed:
Fixed and built on official build 93

:QA Notes:
N/A

:Issues Addressed:
[PLAT-80358] notificationmgr, sensord: are using long deprecated boost
             signals library

Change-Id: I1d8704254c04bc28be063319178e73e3eec63b91